Remove AsanaJson WIP
