+++ package.json
@@ -51,7 +51,6 @@
   },
   "dependencies": {
     "asana": "^0.7.1",
-    "asana-json": "^0.1.13",
     "bluebird": "^2.9.6",
     "react": "^0.12.2",
     "typed-react": "^3.2.0"
