--- src/asana_json.d.ts
@@ -1,40 +1,0 @@
-// Type definitions for asana-api-json
-// Project: https://github.com/asana/asana-api-json/
-// Note: These type definitions are incomplete; only used sections of the API will be defined.
-
-declare module "asana-json" {
-
-  interface Parameter {
-    name: string;
-    type: string;
-    example_values: string[];
-    comment: string;
-    required: boolean;
-  }
-
-  interface Action {
-    name: string;
-    method: string;
-    path: string;
-    comment: string;
-    params?: Parameter[];
-  }
-
-  interface Resource {
-    name: string;
-    comment: string;
-    templates: any[];
-    properties: any[];
-    actions: Action[];
-  }
-
-  export var Attachments: Resource;
-  export var Events: Resource;
-  export var Projects: Resource;
-  export var Stories: Resource;
-  export var Tags: Resource;
-  export var Tasks: Resource;
-  export var Teams: Resource;
-  export var Users: Resource;
-  export var Workspaces: Resource;
-}
