+++ src/resources/templates/resource.ejs
@@ -1,12 +1,12 @@
-/* tslint:disable */
+/// <reference path="../interfaces.ts" />
 
 /**
-* This file is auto-generated by the `asana-api-meta` NodeJS package.
-* We try to keep the generated code pretty clean but there will be lint
-* errors that are just not worth fixing (like unused requires).
-* TODO: maybe we can just disable those specifically and keep this code
-* pretty lint-free too!
-* TODO: Add type information from asana_json.d.ts
-*/
-var json = <%= JSON.stringify(resource, null, 2) %>;
-export = json;
+ * This file is auto-generated by the `asana-api-meta` NodeJS package.
+ * We try to keep the generated code pretty clean but there will be lint
+ * errors that are just not worth fixing (like unused requires).
+ * TODO: maybe we can just disable those specifically and keep this code
+ * pretty lint-free too!
+ */
+/* tslint:disable */
+var resource = <Resource><%= JSON.stringify(resource, null, 2) %>;
+export = resource;
